+++ app/Http/Controllers/AdminController.php
@@ -2,9 +2,11 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\Photo;
 use App\Models\User;
 use Illuminate\Http\Request;
 use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
+use Illuminate\View\View;
 
 class AdminController extends Controller
 {
@@ -20,4 +22,27 @@ class AdminController extends Controller
 
         return view('admin.admin-index', compact('users'));
     }
+
+    public function adminPhotosIndex(): View
+    {
+        // Haal alle foto's op, inclusief soft deleted
+        $photos = Photo::withTrashed()->get();
+
+        return view('admin.photos-index', compact('photos'));
+    }
+
+    public function restorePhoto($id)
+    {
+        // Vind de foto, inclusief soft deleted
+        $photo = Photo::withTrashed()->find($id);
+
+        // Controleer of de foto bestaat
+        if ($photo) {
+            $photo->restore();
+            return redirect()->route('admin.photos-index')->with('status', 'Photo restored successfully!');
+        }
+
+        return redirect()->route('admin.photos-index')->with('error', 'Photo not found.');
+    }
+
 }